% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Gene.R
\name{alternative_transcripts,Gene-method}
\alias{alternative_transcripts,Gene-method}
\title{Get alternative transcripts dataframe}
\usage{
\S4method{alternative_transcripts}{Gene}(object)
}
\arguments{
\item{object}{Object of class \code{Gene} (and derived classes).}
}
\value{
Dataframe with transcript_id and protein_coding (\code{logical})
columns.
}
\description{
Getter for alternative_transcripts attribute of \code{Gene} class
(and derived classes). This attribute can't be modified at construction
time but transcripts can be added/removed only after the creation of the
object by using the dedicated setter.
}
\examples{
# Since Gene is a virtual class, the example is made using a derived
# class.
cg1 <- CodingGene()
alternative_transcripts(cg1)
# The returned row is just a placeholder (it's removed after the first
# modification done by the user).
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
