% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Gene.R
\name{alternative_transcripts}
\alias{alternative_transcripts}
\title{Get Alternative Transcripts}
\usage{
alternative_transcripts(object)
}
\arguments{
\item{object}{Object of class \code{Gene} (and derived classes).}
}
\value{
Dataframe with transcript_id and protein_coding (\code{logical})
columns.
}
\description{
Getter for alternative_transcripts attribute of \code{Gene} class
(and derived classes). This attribute can't be modified at construction
time because transcripts can be added/removed only after the creation of
the object by using the dedicated setter.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
cg1 <- CodingGene()
alternative_transcripts(cg1)
# The returned row is just a placeholder (it's removed after the first
# modification done by the user).
}
\seealso{
\link{alternative_transcripts,Gene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
