% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_tRNAGene.R
\name{amino_acid,tRNAGene-method}
\alias{amino_acid,tRNAGene-method}
\title{Get tRNA Aminoacid}
\usage{
\S4method{amino_acid}{tRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{tRNAGene}.}
}
\value{
Aminoacid sequence (one letter, or empty).
}
\description{
Getter for amino_acid attribute of \code{tRNAGene} class.
}
\examples{
t1 <- tRNAGene(amino_acid = 'K')
amino_acid(t1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
