% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_tRNAGene.R
\name{anticodon<-,tRNAGene-method}
\alias{anticodon<-,tRNAGene-method}
\title{Set tRNA Anticodon}
\usage{
\S4method{anticodon}{tRNAGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{tRNAGene}.}

\item{value}{New value for anticodon.}
}
\value{
The modified object.
}
\description{
Setter for anticodon attribute of \code{tRNAGene} class.
}
\examples{
t1 <- tRNAGene(anticodon = 'GGG')
anticodon(t1)
anticodon(t1) <- 'UUU'
anticodon(t1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
