% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_tRNAGene.R
\name{anticodon,tRNAGene-method}
\alias{anticodon,tRNAGene-method}
\title{Get Anticodon Sequence}
\usage{
\S4method{anticodon}{tRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{tRNAGene}.}
}
\value{
Anticodon sequence.
}
\description{
Getter for anticodon attribute of \code{tRNAGene} class.
}
\examples{
t1 <- tRNAGene(anticodon = 'GGG')
anticodon(t1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
