% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_piRNAGene.R
\name{associated_PIWI_proteins,piRNAGene-method}
\alias{associated_PIWI_proteins,piRNAGene-method}
\title{Get Associated PIWI Proteins}
\usage{
\S4method{associated_PIWI_proteins}{piRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{piRNAGene}.}
}
\value{
List of associated PIWI proteins.
}
\description{
Getter for piRNA_associated_PIWI_proteins attribute of
\code{piRNAGene} class.
}
\examples{
pir1 <- piRNAGene(piRNA_associated_PIWI_proteins=list("PIWIL1","PIWIL2"))
associated_PIWI_proteins(pir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
