% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_piRNAGene.R
\name{associated_PIWI_proteins}
\alias{associated_PIWI_proteins}
\title{Get Associated PIWI Proteins}
\usage{
associated_PIWI_proteins(object)
}
\arguments{
\item{object}{Object of class \code{piRNAGene}.}
}
\value{
List of associated PIWI proteins.
}
\description{
Getter for piRNA_associated_PIWI_proteins attribute of
\code{piRNAGene} class.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
pir1 <- piRNAGene(piRNA_associated_PIWI_proteins=list("PIWIL1","PIWIL2"))
associated_PIWI_proteins(pir1)
}
\seealso{
\link{associated_PIWI_proteins,piRNAGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
