% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_snoRNAGene.R
\name{box_type<-,snoRNAGene-method}
\alias{box_type<-,snoRNAGene-method}
\title{Set snoRNA Box Type}
\usage{
\S4method{box_type}{snoRNAGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{snoRNAGene}.}

\item{value}{New value for snoRNA_box_type.}
}
\value{
The modified object.
}
\description{
Setter for snoRNA_box_type attribute of \code{snoRNAGene} class.
}
\examples{
sno1 <- snoRNAGene(snoRNA_box_type = 'C/D')
box_type(sno1)
box_type(sno1) <- 'H/ACA'
box_type(sno1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
