% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_snoRNAGene.R
\name{box_type,snoRNAGene-method}
\alias{box_type,snoRNAGene-method}
\title{Get snoRNA Box Type}
\usage{
\S4method{box_type}{snoRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{snoRNAGene}.}
}
\value{
snoRNA box type.
}
\description{
Getter for snoRNA_box_type attribute of \code{snoRNAGene} class.
}
\examples{
sno1 <- snoRNAGene(snoRNA_box_type = 'C/D')
box_type(sno1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
