% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_snoRNAGene.R
\name{box_type}
\alias{box_type}
\title{Get snoRNA Box Type}
\usage{
box_type(object)
}
\arguments{
\item{object}{Object of class \code{snoRNAGene}.}
}
\value{
snoRNA box type.
}
\description{
Getter for snoRNA_box_type attribute of \code{snoRNAGene} class.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
sno1 <- snoRNAGene(snoRNA_box_type = 'C/D')
box_type(sno1)
}
\seealso{
\link{box_type,snoRNAGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
