% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CodingGene.R
\name{cds_coordinates,CodingGene-method}
\alias{cds_coordinates,CodingGene-method}
\title{Get CDS Genomic Coordinates}
\usage{
\S4method{cds_coordinates}{CodingGene}(object)
}
\arguments{
\item{object}{Object of class \code{CodingGene}.}
}
\value{
\code{GRanges} object representing CDS coordinates.
}
\description{
Getter for cds_coordinates attribute of \code{CodingGene} class.
}
\examples{
cg1 <- CodingGene(chromosome = 'chr1', strand = '+', cds_start = 50,
cds_end = 90)
cds_coordinates(cg1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
