% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CodingGene.R
\name{cds_coordinates<-,CodingGene-method}
\alias{cds_coordinates<-,CodingGene-method}
\title{Set CDS Genomic Coordinates}
\usage{
\S4method{cds_coordinates}{CodingGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{CodingGene}.}

\item{value}{Named list with chromosome, strand, start and end values
to build the new CDS coordinates.}
}
\value{
The modified object.
}
\description{
Setter for cds_coordinates attribute of \code{CodingGene} class.
}
\examples{
cg1 <- CodingGene(chromosome = 'chr1', strand = '+', cds_start = 40,
cds_end = 80)
cds_coordinates(cg1)
cds_coordinates(cg1) <- list(chromosome = 'chr2', strand = '+', start=30,
end = 50)
cds_coordinates(cg1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
