% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper.R
\name{check_boundaries}
\alias{check_boundaries}
\title{Check Coordinates Respect Boundaries}
\usage{
check_boundaries(grange_in, grange_out, msg = "")
}
\arguments{
\item{grange_in}{\code{GRanges} object.}

\item{grange_out}{\code{GRanges} object.}

\item{msg}{Part of the message to print in case of error.}
}
\value{
\code{TRUE} if coordinates are valid, otherwise the function throws an
error via \code{stop()}.
}
\description{
The function checks that \code{grange_in} coordinates are inside
\code{grange_out} coordinates.
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
