% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper.R
\name{check_coordinates_consistency}
\alias{check_coordinates_consistency}
\title{Check Coordinates Consinstency}
\usage{
check_coordinates_consistency(
    granges_list,
    return_chr = FALSE,
    return_str = FALSE )
}
\arguments{
\item{granges_list}{\code{list} with \code{GRanges} objects to check.}

\item{return_chr}{\code{logical} to return the chromosome or not.}

\item{return_str}{\code{logical} to return the strand or not.}
}
\value{
Chromosome if return_chr is set to TRUE and coordinates are
consistent, strand if return_str is set to TRUE and coordinates are
consistent, TRUE if the coordinates are consistent and the two previous
parameters are set to FALSE, otherwise the function throws an error
via \code{stop()}.
}
\description{
The function checks that the provided \code{GRanges} objects share
both the same chromosome (seqnames) and strand and that chromosome is a
valid human chromosome.
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
