% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper.R
\name{check_ensembl_id}
\alias{check_ensembl_id}
\title{Check Validity of ENSEMBL ID}
\usage{
check_ensembl_id(id, prefix, type, add_msg = "")
}
\arguments{
\item{id}{ID to check.}

\item{prefix}{Expected prefix of \code{id}.}

\item{type}{Part of the message to print in case of error.}

\item{add_msg}{Additional part of the message to print in case of error.}
}
\value{
\code{TRUE} if \code{id} is a valid ENSEMBL ID, otherwise the function
throws an error via \code{stop()}.
}
\description{
The function checks that the provided \code{id} is a valid
ENSEMBL ID.
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
