% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper.R
\name{check_exons_coordinates}
\alias{check_exons_coordinates}
\title{Check Exons Coordinates Validity}
\usage{
check_exons_coordinates(exons, gene)
}
\arguments{
\item{exons}{\code{GRanges} object representing exons coordinates.}

\item{gene}{\code{GRanges} object representing gene coordinates.}
}
\value{
\code{TRUE} if \code{exons} is valid and consistent with \code{gene}, otherwise
the function throws an error via \code{stop()}.
}
\description{
The function checks that the exons coordinates don't overlap
(exons represent the base pair of the genome covered by at least one
transcript of the gene), because if two exons overlap, they are just part
of the same exon. If also gene coordinates are defined, the function
checks that the gene starts and ends with an exon (which are first and last
exons of the gene), because of the definition of gene.
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
