% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper.R
\name{check_in_exons}
\alias{check_in_exons}
\title{Check Coordinates are in Exons}
\usage{
check_in_exons(range_to_check, exons, type)
}
\arguments{
\item{range_to_check}{\code{GRanges} object representing coordinates to check.}

\item{exons}{\code{GRanges} object representing exons coordinates.}

\item{type}{Part of the message to print in case of error.}
}
\value{
\code{TRUE} if coordinates are valid, otherwise the function throws an
error via \code{stop()}.
}
\description{
The function checks that \code{range_to_check} start and end are
inside \code{exons} coordinates range(s).
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
