% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper.R
\name{check_mature_length}
\alias{check_mature_length}
\title{Check Mature Sequence Validity}
\usage{
check_mature_length(object, mature_sequence, short = TRUE)
}
\arguments{
\item{object}{Object of class \code{NonCodingGene}.}

\item{mature_sequence}{\code{character} string representing the slot in which
the mature sequence to check is contained.}

\item{short}{\code{logical} to indicate if \code{object} is of class
SmallNonCodingGene (TRUE) or LongNonCodingGene (FALSE).}
}
\value{
\code{TRUE} if mature sequence is valid, otherwise the function
throws an error via \code{stop()}.
}
\description{
The function checks that the mature sequence length is
compatible with the defined genomic coordinates and gene sequence of
object. Depending on the value of short, the function also checks that
the mature sequence is shorter or longer than 200 nt.
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
