% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper.R
\name{check_slots_length}
\alias{check_slots_length}
\title{Check Length of Each Slot}
\usage{
check_slots_length(obj, class_obj, mother_class_obj, special_slots = "")
}
\arguments{
\item{obj}{Object to check.}

\item{class_obj}{Class of \code{obj}.}

\item{mother_class_obj}{Class from which \code{class_obj} derives.}

\item{special_slots}{Names of the slots NOT to check.}
}
\value{
\code{TRUE} if the slots have a valid length, otherwise the function
throws an error via \code{stop()}.
}
\description{
The function checks that each slot of \code{obj}, except
\code{special_slots}, has a length of 1 (or 0).
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
