% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CodingGene.R
\docType{class}
\name{CodingGene-class}
\alias{CodingGene-class}
\title{CodingGene Class}
\value{
An object of class \code{CodingGene}.
}
\description{
This class represents coding genes and derives from \code{Gene}
class.
}
\section{Slots}{

\describe{
\item{\code{mRNA_ensembl_id}}{\code{character}. mRNA ENSEMBL ID.}

\item{\code{mRNA_coordinates}}{\code{GRanges}. mRNA genomic coordinates.}

\item{\code{cds_coordinates}}{\code{GRanges}. CDS genomic coordinates.}

\item{\code{protein_ensembl_id}}{\code{character}. Protein ENSEMBL ID.}

\item{\code{protein_sequence}}{\code{AAString}. Protein sequence.}

\item{\code{protein_description}}{\code{character}. Protein description.}
}}

\examples{
# An empty CodingGene object.
cg1 <- new("CodingGene")
}
\seealso{
\linkS4class{Gene}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
