% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CodingGene.R
\name{CodingGene-validity}
\alias{CodingGene-validity}
\title{Validity Check for CodingGene class}
\arguments{
\item{object}{Object of class \code{CodingGene}.}
}
\value{
\code{TRUE} if the object is valid, otherwise the function
throws an error via \code{stop()}.
}
\description{
The function checks that: each slot, except the special ones,
is of length 1 (or empty); each ID is a valid ENSEMBL ID; all coordinates
are consistent w.r.t. each other and that gene and protein sequences are
consistent with the provided coordinates.
}
\seealso{
\link{Gene-validity}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
