% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CodingGene.R
\name{CodingGene}
\alias{CodingGene}
\title{Create an Object of Class CodingGene}
\usage{
CodingGene(
    gene_ensembl_id = "ENSG00000000000",
    hugo_symbol = NA_character_,
    gene_complete_name = NA_character_,
    gene_description = NA_character_,
    strand = NA,
    chromosome = NA,
    gene_start = NA,
    gene_end = NA,
    gene_sequence = DNAString(),
    exons_starts = NA,
    exons_ends = NA,
    mRNA_ensembl_id = "ENST00000000000",
    mRNA_start = NA,
    mRNA_end = NA,
    cds_start = NA,
    cds_end = NA,
    protein_ensembl_id = "ENSP00000000000",
    protein_sequence = AAString(),
    protein_description = NA_character_ )
}
\arguments{
\item{gene_ensembl_id}{Gene ENSEMBL ID.}

\item{hugo_symbol}{Gene Hugo symbol.}

\item{gene_complete_name}{Gene complete name.}

\item{gene_description}{Gene description.}

\item{strand}{Strand on which the gene is annotated.}

\item{chromosome}{Chromosome on which the gene is annotated.}

\item{gene_start}{Gene start coordinate.}

\item{gene_end}{Gene end coordinate.}

\item{gene_sequence}{Gene sequence.}

\item{exons_starts}{Exons starts coordinates.}

\item{exons_ends}{Exons ends coordinates.}

\item{mRNA_ensembl_id}{mRNA ENSEMBL ID.}

\item{mRNA_start}{mRNA start coordinate.}

\item{mRNA_end}{mRNA end coordinate.}

\item{cds_start}{CDS start coordinate.}

\item{cds_end}{CDS end coordinate.}

\item{protein_ensembl_id}{Protein ENSEMBL ID.}

\item{protein_sequence}{Protein sequence.}

\item{protein_description}{Protein description.}
}
\value{
New CodingGene object.
}
\description{
Constructor function for \code{CodingGene} class.
}
\examples{
# An empty CodingGene object
cg1 <- CodingGene()
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
