% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper.R
\name{constructor_helper}
\alias{constructor_helper}
\title{Class Constructor Helper Function}
\usage{
constructor_helper(
    chrom,
    str,
    g_st,
    g_en,
    ex_st,
    ex_en,
    m_st = NA,
    m_en = NA,
    c_st = NA,
    c_en = NA,
    pre_st = NA,
    pre_en = NA,
    coding = TRUE )
}
\arguments{
\item{chrom}{Chromosome.}

\item{str}{Strand.}

\item{g_st}{Gene start.}

\item{g_en}{Gene end.}

\item{ex_st}{Exons starts.}

\item{ex_en}{Exons ends.}

\item{m_st}{mRNA start.}

\item{m_en}{mRNA end.}

\item{c_st}{CDS start.}

\item{c_en}{CDS end.}

\item{pre_st}{pre-ncRNA start.}

\item{pre_en}{pre-ncRNA end.}

\item{coding}{\code{logical} to indicate if the function will be used in the
constructor function of an object of class \code{CodingGene} or
\code{NonCodingGene}.}
}
\value{
list containing the new GRanges objects if ex_st and ex_en
are of the same length, otherwise the function throws an error
via \code{stop()}.
}
\description{
The function, depending on the value of \code{coding}, generates
a \code{list} with the necessary \code{GRanges} objects.
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
