% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper.R
\name{create_new_coordinates}
\alias{create_new_coordinates}
\title{Create a \code{GRanges} Object}
\usage{
create_new_coordinates(value, list_names, msg)
}
\arguments{
\item{value}{\code{list} with the necessary components to build
the \code{GRanges} object.}

\item{list_names}{The names that should be found inside \code{value}.}

\item{msg}{Part of the message to print in case of error.}
}
\value{
The new \code{GRanges} object if names(\code{value}) are consistent with
names in \code{list_names}, otherwise the function throws an error via
\code{stop()}.
}
\description{
The function creates a \code{GRanges} object given in input
chromosome, strand, start and end (contained in \code{value}).
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
