% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_HousekeepingNonCodingGene.R
\name{essentiality_score,HousekeepingNonCodingGene-method}
\alias{essentiality_score,HousekeepingNonCodingGene-method}
\title{Get Gene Essentiality Score}
\usage{
\S4method{essentiality_score}{HousekeepingNonCodingGene}(object)
}
\arguments{
\item{object}{Object of class \code{HousekeepingNonCodingGene}
(and derived classes).}
}
\value{
Value of essentiality_score.
}
\description{
Getter for essentiality_score attribute of
\code{HousekeepingNonCodingGene} class (and derived classes).
}
\examples{
# Since HousekeepingNonCodingGene is a virtual class, the example is made
# using a derived class
t1 <- tRNAGene(essentiality_score = 12)
essentiality_score(t1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
