% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_HousekeepingNonCodingGene.R
\name{essentiality_score<-}
\alias{essentiality_score<-}
\title{Set Gene Essentiality Score}
\usage{
essentiality_score(object) <- value
}
\arguments{
\item{object}{Object of class \code{HousekeepingNonCodingGene}
(and derived classes).}

\item{value}{New value for essentiality_score.}
}
\value{
The modified object.
}
\description{
Setter for essentiality_score attribute of
\code{HousekeepingNonCodingGene} class (and derived classes).
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
t1 <- tRNAGene(essentiality_score = 12)
essentiality_score(t1)
essentiality_score(t1) <- 20
essentiality_score(t1)
}
\seealso{
\link{essentiality_score<-,HousekeepingNonCodingGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
