% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Gene.R
\name{exons_coordinates,Gene-method}
\alias{exons_coordinates,Gene-method}
\title{Get Exons Genomic Coordinates}
\usage{
\S4method{exons_coordinates}{Gene}(object)
}
\arguments{
\item{object}{Object of class \code{Gene} (and derived classes).}
}
\value{
\code{GRanges} object representing exons coordinates.
}
\description{
Getter for exons_coordinates attribute of \code{Gene} class
(and derived classes).
}
\examples{
# Since Gene is a virtual class, the example is made using a derived
# class.
cg1 <- CodingGene(chromosome = 'chr1', strand = '+',
exons_starts = c(10,50), exons_ends = c(30, 90))
exons_coordinates(cg1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
