% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Gene.R
\name{exons_coordinates<-}
\alias{exons_coordinates<-}
\title{Set Exons Genomic Coordinates}
\usage{
exons_coordinates(object) <- value
}
\arguments{
\item{object}{Object of class \code{Gene} (and derived classes).}

\item{value}{Named list with chromosome, strand, starts and ends values
to build the new exons coordinates.}
}
\value{
The modified object.
}
\description{
Setter for exons_coordinates attribute of \code{Gene} class
(and derived classes).
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
cg1 <- CodingGene(chromosome = 'chr1', strand = '+',
exons_starts = c(20,60), exons_ends = c(30,70))
exons_coordinates(cg1)
exons_coordinates(cg1) <- list(chromosome = 'chr2', strand = '+',
starts = c(90,150), ends = c(100,400))
exons_coordinates(cg1)
}
\seealso{
\link{exons_coordinates<-,Gene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
