% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Gene.R
\name{exons_coordinates}
\alias{exons_coordinates}
\title{Get Exons Genomic Coordinates}
\usage{
exons_coordinates(object)
}
\arguments{
\item{object}{Object of class \code{Gene} (and derived classes).}
}
\value{
\code{GRanges} object representing exons coordinates.
}
\description{
Getter for exons_coordinates attribute of \code{Gene} class
(and derived classes).
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
cg1 <- CodingGene(chromosome = 'chr1', strand = '+',
exons_starts = c(10,50), exons_ends = c(30, 90))
exons_coordinates(cg1)
}
\seealso{
\link{exons_coordinates,Gene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
