% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_LongNonCodingGene.R
\name{functional_class,LongNonCodingGene-method}
\alias{functional_class,LongNonCodingGene-method}
\title{Get lncRNA Functional class}
\usage{
\S4method{functional_class}{LongNonCodingGene}(object)
}
\arguments{
\item{object}{Object of class \code{LongNonCodingGene}.}
}
\value{
lncRNA functional class.
}
\description{
Getter for lncRNA_functional_class attribute of
\code{LongNonCodingGene} class.
}
\examples{
lnc1 <- LongNonCodingGene(lncRNA_functional_class = 'decoy')
functional_class(lnc1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
