% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_LongNonCodingGene.R
\name{functional_class}
\alias{functional_class}
\title{Get lncRNA Functional Class}
\usage{
functional_class(object)
}
\arguments{
\item{object}{Object of class \code{LongNonCodingGene}.}
}
\value{
lncRNA functional class.
}
\description{
Getter for lncRNA_functional_class attribute of
\code{LongNonCodingGene} class.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
lnc1 <- LongNonCodingGene(lncRNA_functional_class = 'decoy')
functional_class(lnc1)
}
\seealso{
\link{functional_class,LongNonCodingGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
