% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Gene.R
\docType{class}
\name{Gene-class}
\alias{Gene-class}
\title{Gene Virtual Class}
\value{
No objects can be created directly from this class since
it's virtual.
}
\description{
This virtual class is the one from which all the others
derives.
}
\section{Slots}{

\describe{
\item{\code{gene_ensembl_id}}{\code{character}. Gene ENSEMBL ID.}

\item{\code{hugo_symbol}}{\code{character}. Gene Hugo symbol.}

\item{\code{gene_complete_name}}{\code{character}. Gene complete name.}

\item{\code{gene_description}}{\code{character}. Gene description.}

\item{\code{gene_coordinates}}{\code{GRanges}. Gene genomic coordinates.}

\item{\code{gene_sequence}}{\code{DNAString}. Gene sequence.}

\item{\code{exons_coordinates}}{\code{GRanges}. Exons genomic coordinates.}

\item{\code{alternative_transcripts}}{\code{data.frame}.
Alternative transcripts.}
}}

\examples{
# The Gene class is virtual, so no objects can be created directly
# from this class. It needs to be extended by other classes.
}
\seealso{
\linkS4class{CodingGene}, \linkS4class{NonCodingGene}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
