% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Gene.R
\name{gene_complete_name<-,Gene-method}
\alias{gene_complete_name<-,Gene-method}
\title{Set Gene Complete Name}
\usage{
\S4method{gene_complete_name}{Gene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{Gene} (and derived classes).}

\item{value}{New value for gene_complete_name.}
}
\value{
The modified object.
}
\description{
Setter for gene_complete_name attribute of \code{Gene} class
(and derived classes).
}
\examples{
# Since Gene is a virtual class, the example is made using a derived
# class.
cg1 <- CodingGene(gene_complete_name = 'Tumor Protein 53')
gene_complete_name(cg1)
gene_complete_name(cg1) <- 'Breast cancer type 1 susceptibility protein'
gene_complete_name(cg1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
