% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Gene.R
\name{gene_coordinates}
\alias{gene_coordinates}
\title{Get Gene Genomic Coordinates}
\usage{
gene_coordinates(object)
}
\arguments{
\item{object}{Object of class \code{Gene} (and derived classes).}
}
\value{
\code{GRanges} object representing gene coordinates.
}
\description{
Getter for gene_coordinates attribute of \code{Gene} class
(and derived classes).
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
cg1 <- CodingGene(chromosome = 'chr1', strand = '+', gene_start = 210,
gene_end = 800)
gene_coordinates(cg1)
}
\seealso{
\link{gene_coordinates,Gene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
