% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Gene.R
\name{gene_description,Gene-method}
\alias{gene_description,Gene-method}
\title{Get Gene Description}
\usage{
\S4method{gene_description}{Gene}(object)
}
\arguments{
\item{object}{Object of class \code{Gene} (and derived classes).}
}
\value{
Value of gene_description.
}
\description{
Getter for gene_description attribute of \code{Gene} class
(and derived classes).
}
\examples{
# Since Gene is a virtual class, the example is made using a derived
# class.
cg1 <- CodingGene(gene_description = 'TP53 is a tumor suppressor.')
gene_description(cg1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
