% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Gene.R
\name{gene_description}
\alias{gene_description}
\title{Get Gene Description}
\usage{
gene_description(object)
}
\arguments{
\item{object}{Object of class \code{Gene} (and derived classes).}
}
\value{
Value of gene_description.
}
\description{
Getter for gene_description attribute of \code{Gene} class
(and derived classes).
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
cg1 <- CodingGene(gene_description = 'TP53 is a tumor suppressor.')
gene_description(cg1)
}
\seealso{
\link{gene_description,Gene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
