% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Gene.R
\name{gene_id,Gene-method}
\alias{gene_id,Gene-method}
\title{Get Gene ENSEMBL ID}
\usage{
\S4method{gene_id}{Gene}(object)
}
\arguments{
\item{object}{Object of class \code{Gene} (and derived classes).}
}
\value{
Value of gene_ensembl_id.
}
\description{
Getter for gene_ensembl_id attribute of \code{Gene} class
(and derived classes).
}
\examples{
# Since Gene is a virtual class, the example is made using a derived
# class.
cg1 <- CodingGene(gene_ensembl_id = 'ENSG12312312300')
gene_id(cg1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
