% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Gene.R
\name{gene_sequence,Gene-method}
\alias{gene_sequence,Gene-method}
\title{Get Gene Sequence}
\usage{
\S4method{gene_sequence}{Gene}(object)
}
\arguments{
\item{object}{Object of class \code{Gene} (and derived classes).}
}
\value{
Gene sequence.
}
\description{
Getter for gene_sequence attribute of \code{Gene} class
(and derived classes).
}
\examples{
# Since Gene is a virtual class, the example is made using a derived
# class.
cg1 <- CodingGene(gene_sequence = 'ATCATCATCACCATC')
gene_sequence(cg1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
