% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_HousekeepingNonCodingGene.R
\docType{class}
\name{HousekeepingNonCodingGene-class}
\alias{HousekeepingNonCodingGene-class}
\title{HousekeepingNonCodingGene Virtual Class}
\value{
No objects can be created directly from this class since
it's virtual.
}
\description{
This virtual class is the class from which
tRNAGene and rRNAGene classes are derived.
}
\section{Slots}{

\describe{
\item{\code{essentiality_score}}{\code{numeric}. Gene essentiality score.}

\item{\code{ubiquitous_expression}}{\code{logical}.
Gene ubiquitously expressed or not.}
}}

\examples{
# The HousekeepingNonCodingGene class is virtual, so no objects can
# be created directly from this class. It needs to be extended by
# other classes.
}
\seealso{
\linkS4class{NonCodingGene}, \linkS4class{tRNAGene},
\linkS4class{rRNAGene}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
