% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_HousekeepingNonCodingGene.R
\name{HousekeepingNonCodingGene-validity}
\alias{HousekeepingNonCodingGene-validity}
\title{Validity Check for HousekeepingNonCodingGene Virtual Class}
\arguments{
\item{object}{Object of class \code{HousekeepingNonCodingGene}
(and derived classes).}
}
\value{
\code{TRUE} if the object is valid, otherwise the function
throws an error via \code{stop()}.
}
\description{
The function checks that: each slot, except the special ones,
is of length 1 (or empty).
}
\seealso{
\link{NonCodingGene-validity}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
