% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Gene.R
\name{hugo_symbol,Gene-method}
\alias{hugo_symbol,Gene-method}
\title{Get Gene Hugo symbol}
\usage{
\S4method{hugo_symbol}{Gene}(object)
}
\arguments{
\item{object}{Object of class \code{Gene} (and derived classes).}
}
\value{
Value of hugo_symbol.
}
\description{
Getter for hugo_symbol attribute of \code{Gene} class
(and derived classes).
}
\examples{
# Since Gene is a virtual class, the example is made using a derived
# class.
cg1 <- CodingGene(hugo_symbol = 'TP53')
hugo_symbol(cg1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
