% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Gene.R
\name{hugo_symbol<-}
\alias{hugo_symbol<-}
\title{Set Gene Hugo symbol}
\usage{
hugo_symbol(object) <- value
}
\arguments{
\item{object}{Object of class \code{Gene} (and derived classes).}

\item{value}{New value for hugo_symbol.}
}
\value{
The modified object.
}
\description{
Setter for hugo_symbol attribute of \code{Gene} class
(and derived classes).
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
cg1 <- CodingGene(hugo_symbol = 'TP53')
hugo_symbol(cg1)
hugo_symbol(cg1) <- 'BRCA1'
hugo_symbol(cg1)
}
\seealso{
\link{hugo_symbol<-,Gene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
