% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_lengthProduct.R
\name{lengthProduct,CodingGene-method}
\alias{lengthProduct,CodingGene-method}
\title{Get Protein Product Length}
\usage{
\S4method{lengthProduct}{CodingGene}(object)
}
\arguments{
\item{object}{Object of class \code{CodingGene}.}
}
\value{
Protein product length.
}
\description{
Compute length of protein_sequence attribute of \code{CodingGene}
class.
}
\examples{
cg1 <- CodingGene(protein_sequence = 'KRTA')
lengthProduct(cg1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
