% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_lengthProduct.R
\name{lengthProduct,LongNonCodingGene-method}
\alias{lengthProduct,LongNonCodingGene-method}
\title{Get Mature lncRNA Length}
\usage{
\S4method{lengthProduct}{LongNonCodingGene}(object)
}
\arguments{
\item{object}{Object of class \code{LongNonCodingGene}.}
}
\value{
Mature lncRNA length.
}
\description{
Compute length of lncRNA_mature_sequence attribute of
\code{LongNonCodingGene} class.
}
\examples{
lnc1 <- LongNonCodingGene(lncRNA_mature_sequence = paste(rep('AUUUA', 50),
collapse = ''))
lengthProduct(lnc1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
