% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_lengthProduct.R
\name{lengthProduct,miRNAGene-method}
\alias{lengthProduct,miRNAGene-method}
\title{Get Mature miRNA Length}
\usage{
\S4method{lengthProduct}{miRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{miRNAGene}.}
}
\value{
Mature miRNA length.
}
\description{
Compute length of miRNA_mature_sequence attribute of
\code{miRNAGene} class.
}
\examples{
mir1 <- miRNAGene(miRNA_mature_sequence = 'UAUUAUAUAA')
lengthProduct(mir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
