% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_lengthProduct.R
\name{lengthProduct,piRNAGene-method}
\alias{lengthProduct,piRNAGene-method}
\title{Get Mature piRNA Length}
\usage{
\S4method{lengthProduct}{piRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{piRNAGene}.}
}
\value{
Mature piRNA length.
}
\description{
Compute length of piRNA_mature_sequence attribute of
\code{piRNAGene} class.
}
\examples{
pir1 <- piRNAGene(piRNA_mature_sequence = 'UAUUAUAUAA')
lengthProduct(pir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
