% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_lengthProduct.R
\name{lengthProduct,rRNAGene-method}
\alias{lengthProduct,rRNAGene-method}
\title{Get Mature rRNA Length}
\usage{
\S4method{lengthProduct}{rRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{rRNAGene}.}
}
\value{
Mature rRNA length.
}
\description{
Compute length of rRNA_mature_sequence attribute of \code{rRNAGene}
class.
}
\examples{
r1 <- rRNAGene(rRNA_mature_sequence = 'UAUUAUAUAA')
lengthProduct(r1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
