% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_lengthProduct.R
\name{lengthProduct,siRNAGene-method}
\alias{lengthProduct,siRNAGene-method}
\title{Get Mature siRNA Length}
\usage{
\S4method{lengthProduct}{siRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{siRNAGene}.}
}
\value{
Mature siRNA length.
}
\description{
Compute length of siRNA_mature_sequence attribute of
\code{siRNAGene} class.
}
\examples{
sir1 <- siRNAGene(siRNA_mature_sequence = 'UAUUAUAUAA')
lengthProduct(sir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
