% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_lengthProduct.R
\name{lengthProduct,snoRNAGene-method}
\alias{lengthProduct,snoRNAGene-method}
\title{Get Mature snoRNA Length}
\usage{
\S4method{lengthProduct}{snoRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{snoRNAGene}.}
}
\value{
Mature snoRNA length.
}
\description{
Compute length of snoRNA_mature_sequence attribute of
\code{snoRNAGene} class.
}
\examples{
sno1 <- snoRNAGene(snoRNA_mature_sequence = 'UAUUAUAUAA')
lengthProduct(sno1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
