% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_LongNonCodingGene.R
\name{location_class,LongNonCodingGene-method}
\alias{location_class,LongNonCodingGene-method}
\title{Get lncRNA Location Class}
\usage{
\S4method{location_class}{LongNonCodingGene}(object)
}
\arguments{
\item{object}{Object of class \code{LongNonCodingGene}.}
}
\value{
lncRNA location class.
}
\description{
Getter for lncRNA_location_class attribute of
\code{LongNonCodingGene} class.
}
\examples{
lnc1 <- LongNonCodingGene(lncRNA_location_class = 'intergenic')
location_class(lnc1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
