% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_LongNonCodingGene.R
\name{location_class<-,LongNonCodingGene-method}
\alias{location_class<-,LongNonCodingGene-method}
\title{Set lncRNA Location Class}
\usage{
\S4method{location_class}{LongNonCodingGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{LongNonCodingGene}.}

\item{value}{New value for lncRNA_location_class}
}
\value{
The modified object.
}
\description{
Setter for lncRNA_location_class attribute of
LongNonCodingGene` class.
}
\examples{
lnc1 <- LongNonCodingGene(lncRNA_location_class = 'intergenic')
location_class(lnc1)
location_class(lnc1) <- 'bidirectional'
location_class(lnc1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
