% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_LongNonCodingGene.R
\name{location_class<-}
\alias{location_class<-}
\title{Set lncRNA Location Class}
\usage{
location_class(object) <- value
}
\arguments{
\item{object}{Object of class \code{LongNonCodingGene}.}

\item{value}{New value for lncRNA_location_class}
}
\value{
The modified object.
}
\description{
Setter for lncRNA_location_class attribute of
\code{LongNonCodingGene} class.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
lnc1 <- LongNonCodingGene(lncRNA_location_class = 'intergenic')
location_class(lnc1)
location_class(lnc1) <- 'bidirectional'
location_class(lnc1)
}
\seealso{
\link{location_class<-,LongNonCodingGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
