% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_LongNonCodingGene.R
\docType{class}
\name{LongNonCodingGene-class}
\alias{LongNonCodingGene-class}
\title{LongNonCodingGene Class}
\value{
An object of class \code{LongNonCodingGene}.
}
\description{
Class to represent genes encoding for long non-coding RNAs.
}
\section{Slots}{

\describe{
\item{\code{lncRNA_mature_sequence}}{\code{RNAString}. lncRNA mature sequence.}

\item{\code{lncRNA_location_class}}{\code{character}.
lncRNA genomic location class.}

\item{\code{lncRNA_functional_class}}{\code{character}.
lncRNA functional class.}
}}

\examples{
# An empty LongNonCodingGene object
lnc1 <- new("LongNonCodingGene")
}
\seealso{
\linkS4class{RegulatoryNonCodingGene}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
