% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_LongNonCodingGene.R
\name{LongNonCodingGene-validity}
\alias{LongNonCodingGene-validity}
\title{Validity Check for LongNonCodingGene Class}
\arguments{
\item{object}{Object of class \code{LongNonCodingGene}.}
}
\value{
\code{TRUE} if the object is valid, otherwise the function
throws an error via \code{stop()}.
}
\description{
The function checks that: each slots has a length of 1 (or
empty), the inserted location and functional classes are valid and the
mature sequence is consistent with the gene sequence and the defined
coordinates. The mature sequence, being a lncRNA, needs to be longer than
200 nt.
}
\seealso{
\link{RegulatoryNonCodingGene-validity}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
