% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_LongNonCodingGene.R
\name{mature_sequence,LongNonCodingGene-method}
\alias{mature_sequence,LongNonCodingGene-method}
\title{Get Mature lncRNA Sequence}
\usage{
\S4method{mature_sequence}{LongNonCodingGene}(object)
}
\arguments{
\item{object}{Object of class \code{LongNonCodingGene}.}
}
\value{
Mature lncRNA sequence.
}
\description{
Getter forlncRNA_mature_sequence attribute of
\code{LongNonCodingGene} class.
}
\examples{
lnc1 <- LongNonCodingGene(lncRNA_mature_sequence = paste(rep('AUGCA', 50),
collapse = ''))
mature_sequence(lnc1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
