% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_miRNAGene.R
\name{mature_sequence,miRNAGene-method}
\alias{mature_sequence,miRNAGene-method}
\title{Get Mature miRNA Sequence}
\usage{
\S4method{mature_sequence}{miRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{miRNAGene}.}
}
\value{
Mature miRNA sequence.
}
\description{
Getter for miRNA_mature_sequence attribute of \code{miRNAGene}
class.
}
\examples{
mir1 <- miRNAGene(miRNA_mature_sequence = 'AUGCA')
mature_sequence(mir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
