% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_rRNAGene.R
\name{mature_sequence,rRNAGene-method}
\alias{mature_sequence,rRNAGene-method}
\title{Get Mature rRNA Sequence}
\usage{
\S4method{mature_sequence}{rRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{rRNAGene}.}
}
\value{
Mature rRNA sequence.
}
\description{
Getter for rRNA_mature_sequence attribute of \code{rRNAGene} class.
}
\examples{
r1 <- rRNAGene(rRNA_mature_sequence = 'AUGCA')
mature_sequence(r1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
