% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_LongNonCodingGene.R
\name{mature_sequence<-,LongNonCodingGene-method}
\alias{mature_sequence<-,LongNonCodingGene-method}
\title{Set Mature lncRNA Sequence}
\usage{
\S4method{mature_sequence}{LongNonCodingGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{LongNonCodingGene}.}

\item{value}{New value for lncRNA_mature_sequence.}
}
\value{
The modified object.
}
\description{
Setter for lncRNA_mature_sequence attribute of
\code{LongNonCodingGene} class.
}
\examples{
lnc1 <- LongNonCodingGene(lncRNA_mature_sequence = paste(rep('AUGCA', 50),
collapse = ''))
mature_sequence(lnc1)
mature_sequence(lnc1) <- paste(rep('AUUUA', 50), collapse = '')
mature_sequence(lnc1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
