% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_miRNAGene.R
\name{mature_sequence<-,miRNAGene-method}
\alias{mature_sequence<-,miRNAGene-method}
\title{Set Mature miRNA Sequence}
\usage{
\S4method{mature_sequence}{miRNAGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{miRNAGene}.}

\item{value}{New value for miRNA_mature_sequence.}
}
\value{
The modified object.
}
\description{
Setter for miRNA_mature_sequence attribute of \code{miRNAGene}
class.
}
\examples{
mir1 <- miRNAGene(miRNA_mature_sequence = 'AUUA')
mature_sequence(mir1)
mature_sequence(mir1) <- 'AAAGCA'
mature_sequence(mir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
