% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_piRNAGene.R
\name{mature_sequence<-,piRNAGene-method}
\alias{mature_sequence<-,piRNAGene-method}
\title{Set Mature piRNA Sequence}
\usage{
\S4method{mature_sequence}{piRNAGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{piRNAGene}.}

\item{value}{New value for piRNA_mature_sequence.}
}
\value{
The modified object.
}
\description{
Setter for piRNA_mature_sequence attribute of \code{piRNAGene}
class.
}
\examples{
pir1 <- piRNAGene(piRNA_mature_sequence = 'AUUA')
mature_sequence(pir1)
mature_sequence(pir1) <- 'AAAGCA'
mature_sequence(pir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
