% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_siRNAGene.R
\name{mature_sequence<-,siRNAGene-method}
\alias{mature_sequence<-,siRNAGene-method}
\title{Set Mature siRNA Sequence}
\usage{
\S4method{mature_sequence}{siRNAGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{siRNAGene}.}

\item{value}{New value for siRNA_mature_sequence.}
}
\value{
The modified object.
}
\description{
Setter for siRNA_mature_sequence attribute of \code{siRNAGene}
class.
}
\examples{
sir1 <- siRNAGene(siRNA_mature_sequence = 'AUUA')
mature_sequence(sir1)
mature_sequence(sir1) <- 'AAAGCA'
mature_sequence(sir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
