% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_snoRNAGene.R
\name{mature_sequence<-,snoRNAGene-method}
\alias{mature_sequence<-,snoRNAGene-method}
\title{Set Mature snoRNA Sequence}
\usage{
\S4method{mature_sequence}{snoRNAGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{snoRNAGene}.}

\item{value}{New value for snoRNA_mature_sequence.}
}
\value{
The modified object.
}
\description{
Setter for snoRNA_mature_sequence attribute of
\code{snoRNAGene} class.
}
\examples{
sno1 <- snoRNAGene(snoRNA_mature_sequence = 'AUUA')
mature_sequence(sno1)
mature_sequence(sno1) <- 'AAAGCA'
mature_sequence(sno1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
