% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_snRNAGene.R
\name{mature_sequence<-,snRNAGene-method}
\alias{mature_sequence<-,snRNAGene-method}
\title{Set Mature snRNA Sequence}
\usage{
\S4method{mature_sequence}{snRNAGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{snRNAGene}.}

\item{value}{New value for snRNA_mature_sequence.}
}
\value{
The modified object.
}
\description{
Setter for snRNA_mature_sequence attribute of \code{snRNAGene}
class.
}
\examples{
sn1 <- snRNAGene(snRNA_mature_sequence = 'AUUA')
mature_sequence(sn1)
mature_sequence(sn1) <- 'AAAGCA'
mature_sequence(sn1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
