% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_tRNAGene.R
\name{mature_sequence<-,tRNAGene-method}
\alias{mature_sequence<-,tRNAGene-method}
\title{Set Mature tRNA Sequence}
\usage{
\S4method{mature_sequence}{tRNAGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{tRNAGene}.}

\item{value}{New value for tRNA_mature_sequence.}
}
\value{
The modified object.
}
\description{
Setter for tRNA_mature_sequence attribute of \code{tRNAGene} class.
}
\examples{
t1 <- tRNAGene(tRNA_mature_sequence = 'AUUA')
mature_sequence(t1)
mature_sequence(t1) <- 'AAAGCA'
mature_sequence(t1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
