% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_tRNAGene.R
\name{mature_sequence<-}
\alias{mature_sequence<-}
\title{Set Mature Sequence}
\usage{
mature_sequence(object) <- value
}
\arguments{
\item{object}{Object of a class derived from \code{NonCodingGene}.
The specific type of object depends on the method implementation.}

\item{value}{New value for mature sequence.}
}
\value{
The modified object.
}
\description{
Setter for transcript mature sequence.
The specific attribute depends on the method implementation.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
# Example on a tRNAGene object
t1 <- tRNAGene(tRNA_mature_sequence = 'AUUA')
mature_sequence(t1)
mature_sequence(t1) <- 'AAAGCA'
mature_sequence(t1)
}
\seealso{
\link{mature_sequence<-,tRNAGene-method},
\link{mature_sequence<-,rRNAGene-method},
\link{mature_sequence<-,miRNAGene-method},
\link{mature_sequence<-,siRNAGene-method},
\link{mature_sequence<-,piRNAGene-method},
\link{mature_sequence<-,snRNAGene-method},
\link{mature_sequence<-,snoRNAGene-method},
\link{mature_sequence<-,LongNonCodingGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
