% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_siRNAGene.R
\name{mature_sequence,siRNAGene-method}
\alias{mature_sequence,siRNAGene-method}
\title{Get Mature siRNA sequence}
\usage{
\S4method{mature_sequence}{siRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{siRNAGene}.}
}
\value{
Mature siRNA sequence.
}
\description{
Getter for siRNA_mature_sequence attribute of \code{siRNAGene}
class.
}
\examples{
sir1 <- siRNAGene(siRNA_mature_sequence = 'AUGCA')
mature_sequence(sir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
