% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_snoRNAGene.R
\name{mature_sequence,snoRNAGene-method}
\alias{mature_sequence,snoRNAGene-method}
\title{Get Mature snoRNA Sequence}
\usage{
\S4method{mature_sequence}{snoRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{snoRNAGene}.}
}
\value{
Mature snoRNA sequence.
}
\description{
Getter for snoRNA_mature_sequence attribute of \code{snoRNAGene}
class.
}
\examples{
sno1 <- snoRNAGene(snoRNA_mature_sequence = 'AUGCA')
mature_sequence(sno1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
