% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_snRNAGene.R
\name{mature_sequence,snRNAGene-method}
\alias{mature_sequence,snRNAGene-method}
\title{Get Mature snRNA Sequence}
\usage{
\S4method{mature_sequence}{snRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{snRNAGene}.}
}
\value{
Mature snRNA sequence.
}
\description{
Getter for snRNA_mature_sequence attribute of \code{snRNAGene}
class.
}
\examples{
sn1 <- snRNAGene(snRNA_mature_sequence = 'AUGCA')
mature_sequence(sn1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
