% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_tRNAGene.R
\name{mature_sequence,tRNAGene-method}
\alias{mature_sequence,tRNAGene-method}
\title{Get Mature tRNA Sequence}
\usage{
\S4method{mature_sequence}{tRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{tRNAGene}.}
}
\value{
Mature tRNA sequence.
}
\description{
Getter for tRNA_mature_sequence attribute of \code{tRNAGene} class.
}
\examples{
t1 <- tRNAGene(tRNA_mature_sequence = 'AUGCA')
mature_sequence(t1)
}
