% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_miRNAGene.R
\docType{class}
\name{miRNAGene-class}
\alias{miRNAGene-class}
\title{miRNAGene Class}
\value{
An object of class \code{miRNAGene}.
}
\description{
Class to represent genes encoding for micro RNAs.
}
\section{Slots}{

\describe{
\item{\code{miRNA_mature_sequence}}{\code{RNAString}. miRNA mature sequence.}

\item{\code{seed_sequence}}{\code{RNAString}. miRNA seed sequence.}
}}

\examples{
# An empty miRNAGene object
mir1 <- new("miRNAGene")
}
\seealso{
\linkS4class{RegulatoryNonCodingGene}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
