% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_miRNAGene.R
\name{miRNAGene-validity}
\alias{miRNAGene-validity}
\title{Validity Check for miRNAGene Class}
\arguments{
\item{object}{Object of class \code{miRNAGene}.}
}
\value{
\code{TRUE} if the object is valid, otherwise the function
throws an error via \code{stop()}.
}
\description{
The function checks that seed sequence is actually contained
in the mature miRNA sequence and that mature miRNA sequence is consistent
with the gene sequence and coordinates defined.
}
\seealso{
\link{RegulatoryNonCodingGene-validity}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
