% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_miRNAGene.R
\name{miRNAGene}
\alias{miRNAGene}
\title{Create an Object of Class miRNAGene}
\usage{
miRNAGene(
    gene_ensembl_id = "ENSG00000000000",
    hugo_symbol = NA_character_,
    gene_complete_name = NA_character_,
    gene_description = NA_character_,
    strand = NA,
    chromosome = NA,
    gene_start = NA,
    gene_end = NA,
    gene_sequence = DNAString(),
    exons_starts = NA,
    exons_ends = NA,
    pre_ncRNA_ensembl_id = "ENST00000000000",
    pre_ncRNA_start = NA,
    pre_ncRNA_end = NA,
    regulatory_functions = list(),
    targets_ensembl_id = list(),
    miRNA_mature_sequence = RNAString(),
    seed_sequence = RNAString() )
}
\arguments{
\item{gene_ensembl_id}{Gene ENSEMBL ID.}

\item{hugo_symbol}{Gene Hugo symbol.}

\item{gene_complete_name}{Gene complete name.}

\item{gene_description}{Gene description.}

\item{strand}{Strand on which the gene is annotated.}

\item{chromosome}{Chromosome on which the gene is annotated.}

\item{gene_start}{Gene start coordinate.}

\item{gene_end}{Gene end coordinate.}

\item{gene_sequence}{Gene sequence.}

\item{exons_starts}{Exons starts coordinates.}

\item{exons_ends}{Exons ends coordinates.}

\item{pre_ncRNA_ensembl_id}{pre-ncRNA ENSEMBL ID.}

\item{pre_ncRNA_start}{pre-ncRNA start coordinate.}

\item{pre_ncRNA_end}{pre-ncRNA end coordinate.}

\item{regulatory_functions}{Gene regulatory functions.}

\item{targets_ensembl_id}{ENSEMBL IDs of the gene targets.}

\item{miRNA_mature_sequence}{Mature miRNA sequence.}

\item{seed_sequence}{miRNA seed sequence.}
}
\value{
New miRNAGene object.
}
\description{
Constructor function for \code{miRNAGene} class.
}
\examples{
# An empty miRNAGene object
mir1 <- miRNAGene()
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
