% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper.R
\name{modify_list}
\alias{modify_list}
\title{Modify Input List}
\usage{
modify_list(object, slot, action, value, msg)
}
\arguments{
\item{object}{An object of class \code{RegulatoryNonCodingGene}.}

\item{slot}{Slot in which there's the list to modify.}

\item{action}{The action to perform on the list of object in slot.}

\item{value}{The element to append/remove to/from the list of the object
in slot.}

\item{msg}{Part of the message to print in case of error.}
}
\value{
The modified \code{list}.
}
\description{
The function appends/removes the element specified by \code{value}
from the \code{list} attribute of \code{object}.
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
