% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CodingGene.R
\name{mRNA_coordinates<-,CodingGene-method}
\alias{mRNA_coordinates<-,CodingGene-method}
\title{Set mRNA Genomic Coordinates}
\usage{
\S4method{mRNA_coordinates}{CodingGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{CodingGene}.}

\item{value}{Named list with chromosome, strand, start and end values
to build the new mRNA coordinates.}
}
\value{
The modified object.
}
\description{
Setter for mRNA_coordinates attribute of \code{CodingGene} class.
}
\examples{
cg1 <- CodingGene(chromosome = 'chr1', strand = '+', mRNA_start = 40,
mRNA_end = 80)
mRNA_coordinates(cg1)
mRNA_coordinates(cg1) <- list(chromosome = 'chr2', strand = '+',
start = 30, end = 50)
mRNA_coordinates(cg1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
