% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CodingGene.R
\name{mRNA_coordinates}
\alias{mRNA_coordinates}
\title{Get mRNA Genomic Coordinates}
\usage{
mRNA_coordinates(object)
}
\arguments{
\item{object}{Object of class \code{CodingGene}.}
}
\value{
\code{GRanges} object representing mRNA coordinates.
}
\description{
Getter for mRNA_coordinates attribute of \code{CodingGene} class.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
cg1 <- CodingGene(chromosome = 'chr1', strand = '+', mRNA_start = 20,
mRNA_end = 30)
mRNA_coordinates(cg1)
}
\seealso{
\link{mRNA_coordinates,CodingGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
