% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CodingGene.R
\name{mRNA_id,CodingGene-method}
\alias{mRNA_id,CodingGene-method}
\title{Get mRNA ENSEMBL ID}
\usage{
\S4method{mRNA_id}{CodingGene}(object)
}
\arguments{
\item{object}{Object of class \code{CodingGene}.}
}
\value{
Value of mRNA_ensembl_id.
}
\description{
Getter for mRN A_ensembl_id attribute of \code{CodingGene} class.
}
\examples{
cg1 <- CodingGene(mRNA_ensembl_id = 'ENST12341234101')
mRNA_id(cg1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
