% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CodingGene.R
\name{mRNA_id<-}
\alias{mRNA_id<-}
\title{Set mRNA ENSEMBL ID}
\usage{
mRNA_id(object) <- value
}
\arguments{
\item{object}{Object of class \code{CodingGene}.}

\item{value}{New value for mRNA_ensembl_id.}
}
\value{
The modified object.
}
\description{
Setter for mRNA_ensembl_id attribute of \code{CodingGene} class.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
cg1 <- CodingGene(mRNA_ensembl_id = 'ENST12345678944')
mRNA_id(cg1)
mRNA_id(cg1) <- 'ENST11111000004'
mRNA_id(cg1)
}
\seealso{
\link{mRNA_id<-,CodingGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
