% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CodingGene.R
\name{mRNA_id}
\alias{mRNA_id}
\title{Get mRNA ENSEMBL ID}
\usage{
mRNA_id(object)
}
\arguments{
\item{object}{Object of class \code{CodingGene}.}
}
\value{
Value of mRNA_ensembl_id.
}
\description{
Getter for mRNA_ensembl_id attribute of \code{CodingGene} class.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
cg1 <- CodingGene(mRNA_ensembl_id = 'ENST12341234101')
mRNA_id(cg1)
}
\seealso{
\link{mRNA_id,CodingGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
