% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_NonCodingGene.R
\docType{class}
\name{NonCodingGene-class}
\alias{NonCodingGene-class}
\title{NonCodingGene Virtual Class}
\value{
No objects can be created directly from this class since
it's virtual.
}
\description{
This virtual class is the class from which
HousekeepingNonCodingGene and RegulatoryNonCodingGene classes are derived.
}
\section{Slots}{

\describe{
\item{\code{pre_ncRNA_ensembl_id}}{\code{character}. pre-ncRNA ENSEMBL ID.}

\item{\code{pre_ncRNA_coordinates}}{\code{GRanges}.
pre-ncRNA genomic coordinates.}
}}

\examples{
# The NonCodingGene class is virtual, so no objects can be created
# directly from this class. It needs to be extended by other classes.
}
\seealso{
\linkS4class{Gene}, \linkS4class{HousekeepingNonCodingGene},
\linkS4class{RegulatoryNonCodingGene}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
