% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_NonCodingGene.R
\name{NonCodingGene-validity}
\alias{NonCodingGene-validity}
\title{Validity Check for NonCodingGene Virtual Class}
\arguments{
\item{object}{Object of class \code{NonCodingGene} (and derived classes).}
}
\value{
\code{TRUE} if the object is valid, otherwise the function
throws an error via \code{stop()}.
}
\description{
The function checks that: each slot, except the special ones,
is of length 1 (or empty); pre-ncRNA ID is a valid ENSEMBL ID; all
coordinates are consistent w.r.t. each other and that gene sequence is
consistent with the provided coordinates. It also checks that no
protein-coding transcripts are present in alternative_transcripts slot.
}
\examples{
# The NonCodingGene class is virtual, so no objects can be created
# directly from this class. It needs to be extended by other classes.
}
\seealso{
\link{Gene-validity}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
