% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper.R
\name{nonCodingLength}
\alias{nonCodingLength}
\title{Compute Product Length}
\usage{
nonCodingLength(object, slot, name)
}
\arguments{
\item{object}{An object of class \code{NonCodingGene}.}

\item{slot}{Slot of which the length is computed.}

\item{name}{Name assigned to the output.}
}
\value{
Length of attribute \code{slot} in \code{object}.
}
\description{
The function computes the length of attribute \code{slot} in
\code{object}.
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
