% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_siRNAGene.R
\name{off_targets<-,siRNAGene-method}
\alias{off_targets<-,siRNAGene-method}
\title{Add/Remove Off-Targets}
\usage{
\S4method{off_targets}{siRNAGene}(object, action = "add") <- value
}
\arguments{
\item{object}{Object of class \code{siRNAGene}.}

\item{action}{Action to be done: can be 'add' or 'remove'.}

\item{value}{Off-target to add/remove.}
}
\value{
The modified object.
}
\description{
Setter for siRNA_off_targets attribute of
\code{siRNAGene} class. By choosing an action, which can be
'add' or 'remove', it's possible to add and remove, respectively,
off-targets from siRNA_off_targets attribute.
}
\examples{
sir1 <- siRNAGene()
off_targets(sir1)
off_targets(sir1, action = 'add') <- 'IRF2'
off_targets(sir1)
off_targets(sir1, action = 'remove') <- 'IRF2'
off_targets(sir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
