% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_siRNAGene.R
\name{off_targets,siRNAGene-method}
\alias{off_targets,siRNAGene-method}
\title{Get Off-Targets}
\usage{
\S4method{off_targets}{siRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{siRNAGene}.}
}
\value{
List of off-targets.
}
\description{
Getter for siRNA_off_targets attribute of \code{siRNAGene} class.
}
\examples{
sir1 <- siRNAGene(siRNA_off_targets = list("BRCA1","BRCA2"))
off_targets(sir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
