% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_siRNAGene.R
\name{off_targets}
\alias{off_targets}
\title{Get Off-Targets}
\usage{
off_targets(object)
}
\arguments{
\item{object}{Object of class \code{siRNAGene}.}
}
\value{
List of off-targets.
}
\description{
Getter for siRNA_off_targets attribute of \code{siRNAGene} class.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
sir1 <- siRNAGene(siRNA_off_targets = list("BRCA1","BRCA2"))
off_targets(sir1)
}
\seealso{
\link{off_targets,siRNAGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
