% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_piRNAGene.R
\docType{class}
\name{piRNAGene-class}
\alias{piRNAGene-class}
\title{piRNAGene Class}
\value{
An object of class \code{piRNAGene}.
}
\description{
Class to represent genes encoding for piwi-interacting RNAs.
}
\section{Slots}{

\describe{
\item{\code{piRNA_mature_sequence}}{\code{RNAString}. piRNA mature sequence.}

\item{\code{piRNA_associated_PIWI_proteins}}{\code{list}.
PIWI proteins associated to piRNA.}
}}

\examples{
# An empty piRNAGene object
pir1 <- new("piRNAGene")
}
\seealso{
\linkS4class{RegulatoryNonCodingGene}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
