% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_piRNAGene.R
\name{piRNAGene-validity}
\alias{piRNAGene-validity}
\title{Validity Check for piRNAGene Class}
\arguments{
\item{object}{Object of class \code{piRNAGene}.}
}
\value{
\code{TRUE} if the object is valid, otherwise the function
throws an error via \code{stop()}.
}
\description{
The function checks that mature piRNA sequence is consistent
with the gene sequence and coordinates defined.
}
\seealso{
\link{RegulatoryNonCodingGene-validity}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
