% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_PlotGene.R
\name{plot,NonCodingGene,ANY-method}
\alias{plot,NonCodingGene,ANY-method}
\title{Plot NonCodingGene Objects}
\usage{
\S4method{plot}{NonCodingGene,ANY}(x, y = 0)
}
\arguments{
\item{x}{Object of class \code{NonCodingGene} (and derived classes).}

\item{y}{Required parameter to extend the generic function 'plot()'.
Its value doesn't influence the final plot.}
}
\value{
Gene plot.
}
\description{
The function plots the gene's exons and, if provided,
the pre-nc RNA.
}
\examples{
mir1 <- miRNAGene(
    exons_starts = c(15, 40, 90),
    exons_ends = c(20, 50, 100),
    pre_ncRNA_start = 44, pre_ncRNA_end = 95,
    chromosome = 'chr1', strand = '-',
    pre_ncRNA_ensembl_id = "ENST09090911190",
    gene_ensembl_id = "ENSG00033300033")

plot(mir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
