% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_NonCodingGene.R
\name{pre_ncRNA_coordinates,NonCodingGene-method}
\alias{pre_ncRNA_coordinates,NonCodingGene-method}
\title{Get pre-ncRNA Genomic Coordinates}
\usage{
\S4method{pre_ncRNA_coordinates}{NonCodingGene}(object)
}
\arguments{
\item{object}{Object of class \code{NonCodingGene} (and derived classes).}
}
\value{
\code{GRanges} object representing pre-ncRNA coordinates.
}
\description{
Getter for pre_ncRNA_coordinates attribute of \code{NonCodingGene}
class (and derived classes).
}
\examples{
# Since NonCodingGene is a virtual class, the example is made using a
# derived class
mir1 <- miRNAGene(chromosome = 'chr1', strand = '+', pre_ncRNA_start=210,
pre_ncRNA_end = 800)
pre_ncRNA_coordinates(mir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
