% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_NonCodingGene.R
\name{pre_ncRNA_coordinates<-,NonCodingGene-method}
\alias{pre_ncRNA_coordinates<-,NonCodingGene-method}
\title{Set pre-ncRNA Genomic Coordinates}
\usage{
\S4method{pre_ncRNA_coordinates}{NonCodingGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{NonCodingGene} (and derived classes).}

\item{value}{Named list with chromosome, strand, start and end values
to build the new pre-ncRNA coordinates.}
}
\value{
The modified object.
}
\description{
Setter for pre_ncRNA_coordinates attribute of \code{NonCodingGene}
class (and derived classes).
}
\examples{
# Since NonCodingGene is a virtual class, the example is made using a
# derived class
mir1 <- miRNAGene(chromosome = 'chr1', strand = '+', pre_ncRNA_start = 20,
pre_ncRNA_end = 300)
pre_ncRNA_coordinates(mir1)
pre_ncRNA_coordinates(mir1) <- list(chromosome = 'chr2', strand = '+',
start = 30, end = 50)
pre_ncRNA_coordinates(mir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
