% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_NonCodingGene.R
\name{pre_ncRNA_coordinates}
\alias{pre_ncRNA_coordinates}
\title{Get pre-ncRNA Genomic Coordinates}
\usage{
pre_ncRNA_coordinates(object)
}
\arguments{
\item{object}{Object of class \code{NonCodingGene} (and derived classes).}
}
\value{
\code{GRanges} object representing pre-ncRNA coordinates.
}
\description{
Getter for pre_ncRNA_coordinates attribute of \code{NonCodingGene}
class (and derived classes).
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
mir1 <- miRNAGene(chromosome = 'chr1', strand = '+', pre_ncRNA_start=210,
pre_ncRNA_end = 800)
pre_ncRNA_coordinates(mir1)
}
\seealso{
\link{pre_ncRNA_coordinates,NonCodingGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
