% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_NonCodingGene.R
\name{pre_ncRNA_id<-}
\alias{pre_ncRNA_id<-}
\title{Set pre-ncRNA ENSEMBL ID}
\usage{
pre_ncRNA_id(object) <- value
}
\arguments{
\item{object}{Object of class \code{NonCodingGene} (and derived classes).}

\item{value}{New value for pre_ncRNA_ensembl_id.}
}
\value{
The modified object.
}
\description{
Setter for pre_ncRNA_ensembl_id attribute of \code{NonCodingGene}
class (and derived classes).
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
mir1 <- miRNAGene(pre_ncRNA_ensembl_id = 'ENST12312312300')
pre_ncRNA_id(mir1)
pre_ncRNA_id(mir1) <- 'ENST00099988812'
pre_ncRNA_id(mir1)
}
\seealso{
\link{pre_ncRNA_id<-,NonCodingGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
