% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CodingGene.R
\name{protein_id,CodingGene-method}
\alias{protein_id,CodingGene-method}
\title{Get Protein ENSEMBL ID}
\usage{
\S4method{protein_id}{CodingGene}(object)
}
\arguments{
\item{object}{Object of class \code{CodingGene}.}
}
\value{
Protein ENSEMBL ID.
}
\description{
Getter for protein_ensembl_id attribute of \code{CodingGene} class.
}
\examples{
cg1 <- CodingGene(protein_ensembl_id = 'ENSP12312312300')
protein_id(cg1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
