% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CodingGene.R
\name{protein_id<-,CodingGene-method}
\alias{protein_id<-,CodingGene-method}
\title{Set Protein ENSEMBL ID}
\usage{
\S4method{protein_id}{CodingGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{CodingGene}.}

\item{value}{New value for protein_ensembl_id.}
}
\value{
The modified object.
}
\description{
Setter for protein_ensembl_id attribute of \code{CodingGene} class.
}
\examples{
cg1 <- CodingGene(protein_ensembl_id = 'ENSP12345678944')
protein_id(cg1)
protein_id(cg1) <- 'ENSP11111000004'
protein_id(cg1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
