% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CodingGene.R
\name{protein_id}
\alias{protein_id}
\title{Get Protein ENSEMBL ID}
\usage{
protein_id(object)
}
\arguments{
\item{object}{Object of class \code{CodingGene}.}
}
\value{
Protein ENSEMBL ID.
}
\description{
Getter for protein_ensembl_id attribute of \code{CodingGene} class.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
cg1 <- CodingGene(protein_ensembl_id = 'ENSP12312312300')
protein_id(cg1)
}
\seealso{
\link{protein_id,CodingGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
