% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CodingGene.R
\name{protein_sequence,CodingGene-method}
\alias{protein_sequence,CodingGene-method}
\title{Get Protein Sequence}
\usage{
\S4method{protein_sequence}{CodingGene}(object)
}
\arguments{
\item{object}{Object of class \code{CodingGene}.}
}
\value{
Protein sequence.
}
\description{
Getter for protein_sequence attribute of \code{CodingGene} class.
}
\examples{
cg1 <- CodingGene(protein_sequence = 'KRT')
protein_sequence(cg1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
