% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CodingGene.R
\name{protein_sequence}
\alias{protein_sequence}
\title{Get Protein Sequence}
\usage{
protein_sequence(object)
}
\arguments{
\item{object}{Object of class \code{CodingGene}.}
}
\value{
Protein sequence.
}
\description{
Getter for protein_sequence attribute of \code{CodingGene} class.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
cg1 <- CodingGene(protein_sequence = 'KRT')
protein_sequence(cg1)
}
\seealso{
\link{protein_sequence,CodingGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
