% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_RegulatoryNonCodingGene.R
\name{regulatory_functions<-,RegulatoryNonCodingGene-method}
\alias{regulatory_functions<-,RegulatoryNonCodingGene-method}
\title{Add/Remove Regulatory Functions}
\usage{
\S4method{regulatory_functions}{RegulatoryNonCodingGene}(object,
action) <- value
}
\arguments{
\item{object}{Object of class \code{RegulatoryNonCodingGene}
(and derived classes).}

\item{action}{Action to be done: can be 'add' or 'remove'.}

\item{value}{Regulatory function to add/remove.}
}
\value{
The modified object.
}
\description{
Setter for regulatory_functions attribute of
RegulatoryNonCodingGene class (and derived classes). By choosing an
action, which can be 'add' or 'remove', it's possible to add and remove,
respectively, regulatory function from regulatory_functions attribute.
}
\examples{
# Since RegulatoryNonCodingGene is a virtual class, the example is made
# using a derived class
mir1 <- miRNAGene()
regulatory_functions(mir1)
regulatory_functions(mir1, action = 'add') <- 'transcription factor'
regulatory_functions(mir1)
regulatory_functions(mir1, action = 'remove') <- 'transcription factor'
regulatory_functions(mir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
