% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_RegulatoryNonCodingGene.R
\name{regulatory_functions<-}
\alias{regulatory_functions<-}
\title{Add/Remove Regulatory Functions}
\usage{
regulatory_functions(object, action = "add") <- value
}
\arguments{
\item{object}{Object of class \code{RegulatoryNonCodingGene}
(and derived classes).}

\item{action}{Action to be done: can be 'add' or 'remove'.}

\item{value}{Regulatory function to add/remove.}
}
\value{
The modified object.
}
\description{
Setter for regulatory_functions attribute of
\code{RegulatoryNonCodingGene} class (and derived classes). By choosing an
action, which can be 'add' or 'remove', it's possible to add and remove,
respectively, regulatory function from regulatory_functions attribute.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
mir1 <- miRNAGene()
regulatory_functions(mir1)
regulatory_functions(mir1, action = 'add') <- 'transcription factor'
regulatory_functions(mir1)
regulatory_functions(mir1, action = 'remove') <- 'transcription factor'
regulatory_functions(mir1)
}
\seealso{
\link{regulatory_functions<-,RegulatoryNonCodingGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
