% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_RegulatoryNonCodingGene.R
\name{regulatory_functions}
\alias{regulatory_functions}
\title{Get Regulatory Functions}
\usage{
regulatory_functions(object)
}
\arguments{
\item{object}{Object of class \code{RegulatoryNonCodingGene}
(and derived classes).}
}
\value{
List of regulatory functions.
}
\description{
Getter for regulatory_functions attribute of
\code{RegulatoryNonCodingGene} class (and derived classes).
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
mir1 <- miRNAGene(regulatory_functions = list("Trascription modulation"))
regulatory_functions(mir1)
}
\seealso{
\link{regulatory_functions,RegulatoryNonCodingGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
