% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_RegulatoryNonCodingGene.R
\docType{class}
\name{RegulatoryNonCodingGene-class}
\alias{RegulatoryNonCodingGene-class}
\title{RegulatoryNonCodingGene Virtual Class}
\value{
No objects can be created directly from this class since
it's virtual.
}
\description{
This virtual class is the class from which
SmallNonCodingGene and LongNonCodingGene classes are derived.
}
\section{Slots}{

\describe{
\item{\code{regulatory_functions}}{\code{list}.
Regulatory functions of the gene.}

\item{\code{targets_ensembl_id}}{\code{list}. ENSEMBL IDs of the gene targets.}
}}

\examples{
# The RegulatoryNonCodingGene class is virtual, so no objects can
# be created directly from this class. It needs to be extended by
# other classes.
}
\seealso{
\linkS4class{NonCodingGene}, \linkS4class{SmallNonCodingGene},
\linkS4class{LongNonCodingGene}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
