% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_RegulatoryNonCodingGene.R
\name{RegulatoryNonCodingGene-validity}
\alias{RegulatoryNonCodingGene-validity}
\title{Validity Check for RegulatoryNonCodingGene Class}
\arguments{
\item{object}{Object of class \code{RegulatoryNonCodingGene}
(and derived classes).}
}
\value{
\code{TRUE} if the object is valid, otherwise the function
throws an error via \code{stop()}.
}
\description{
The function checks that each ID in targets_ensembl_id slot
is a valid ENSEMBL ID.
}
\seealso{
\link{NonCodingGene-validity}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
