% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_rRNAGene.R
\name{rRNA_type,rRNAGene-method}
\alias{rRNA_type,rRNAGene-method}
\title{Get rRNA Type}
\usage{
\S4method{rRNA_type}{rRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{rRNAGene}.}
}
\value{
rRNA type value.
}
\description{
Getter for rRNA_type attribute of \code{rRNAGene} class.
}
\examples{
r1 <- rRNAGene(rRNA_type = '5S')
rRNA_type(r1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
