% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_rRNAGene.R
\name{rRNA_type<-,rRNAGene-method}
\alias{rRNA_type<-,rRNAGene-method}
\title{Set rRNA Type}
\usage{
\S4method{rRNA_type}{rRNAGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{rRNAGene}.}

\item{value}{New value for rRNA_type}
}
\value{
The modified object.
}
\description{
Setter for rRNA_type attribute of \code{rRNAGene} class.
}
\examples{
r1 <- rRNAGene(rRNA_type = '5S')
rRNA_type(r1)
rRNA_type(r1) <- '18S'
rRNA_type(r1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
