% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_rRNAGene.R
\name{rRNA_type}
\alias{rRNA_type}
\title{Get rRNA Type}
\usage{
rRNA_type(object)
}
\arguments{
\item{object}{Object of class \code{rRNAGene}.}
}
\value{
rRNA type value.
}
\description{
Getter for rRNA_type attribute of \code{rRNAGene} class.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
r1 <- rRNAGene(rRNA_type='5S')
rRNA_type(r1)
}
\seealso{
\link{rRNA_type,rRNAGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
