% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_rRNAGene.R
\docType{class}
\name{rRNAGene-class}
\alias{rRNAGene-class}
\title{rRNAGene Class}
\value{
An object of class \code{rRNAGene}.
}
\description{
Class to represent genes encoding for ribosomal RNAs.
}
\section{Slots}{

\describe{
\item{\code{rRNA_mature_sequence}}{\code{RNAString}. rRNA mature sequence.}

\item{\code{rRNA_type}}{\code{character}. Ribosomal RNA type (e.g. 5S) .}
}}

\examples{
# An empty rRNAGene object
r1 <- new("rRNAGene")
}
\seealso{
\linkS4class{HousekeepingNonCodingGene}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
