% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_rRNAGene.R
\name{rRNAGene-validity}
\alias{rRNAGene-validity}
\title{Validity Check for rRNAGene Class}
\arguments{
\item{object}{Object of class \code{rRNAGene}.}
}
\value{
\code{TRUE} if the object is valid, otherwise the function
throws an error via \code{stop()}.
}
\description{
The function checks that: each slot, except the special ones,
is of length 1 (or empty); maure tRNA sequence is compatible with both the
defined coordinates and the gene sequence and finally, that rRNA type and
chromosome are consistent (e.g. rRNA 5S genes are only on chromosome 1).
}
\seealso{
\link{HousekeepingNonCodingGene-validity}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
