% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_rRNAGene.R
\name{rRNAGene}
\alias{rRNAGene}
\title{Create an Object of Class rRNAGene}
\usage{
rRNAGene(
    gene_ensembl_id = "ENSG00000000000",
    hugo_symbol = NA_character_,
    gene_complete_name = NA_character_,
    gene_description = NA_character_,
    strand = NA,
    chromosome = NA,
    gene_start = NA,
    gene_end = NA,
    gene_sequence = DNAString(),
    exons_starts = NA,
    exons_ends = NA,
    pre_ncRNA_ensembl_id = "ENST00000000000",
    pre_ncRNA_start = NA,
    pre_ncRNA_end = NA,
    essentiality_score = NA_real_,
    ubiquitous_expression = NA,
    rRNA_mature_sequence = RNAString(),
    rRNA_type = NA_character_ )
}
\arguments{
\item{gene_ensembl_id}{Gene ENSEMBL ID.}

\item{hugo_symbol}{Gene Hugo symbol.}

\item{gene_complete_name}{Gene complete name.}

\item{gene_description}{Gene description.}

\item{strand}{Strand on which the gene is annotated.}

\item{chromosome}{Chromosome on which the gene is annotated.}

\item{gene_start}{Gene start coordinate.}

\item{gene_end}{Gene end coordinate.}

\item{gene_sequence}{Gene sequence.}

\item{exons_starts}{Exons starts coordinates.}

\item{exons_ends}{Exons ends coordinates.}

\item{pre_ncRNA_ensembl_id}{pre-ncRNA ENSEMBL ID.}

\item{pre_ncRNA_start}{pre-ncRNA start coordinate.}

\item{pre_ncRNA_end}{pre-ncRNA end coordinate.}

\item{essentiality_score}{Gene essentiality score.}

\item{ubiquitous_expression}{\code{logical} for ubiquitous expression.}

\item{rRNA_mature_sequence}{rRNA mature sequence.}

\item{rRNA_type}{Type of rRNA.}
}
\value{
New rRNAGene object.
}
\description{
Constructor function for \code{rRNAGene} class.
}
\examples{
# An empty rRNAGene object
r1 <- rRNAGene()
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
