% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_miRNAGene.R
\name{seed_sequence,miRNAGene-method}
\alias{seed_sequence,miRNAGene-method}
\title{Get miRNA Seed Sequence}
\usage{
\S4method{seed_sequence}{miRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{miRNAGene}.}
}
\value{
Mature miRNA sequence.
}
\description{
Getter for seed_sequence attribute of \code{miRNAGene} class.
}
\examples{
mir1 <- miRNAGene(seed_sequence = 'AUGCA')
seed_sequence(mir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
