% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_miRNAGene.R
\name{seed_sequence<-}
\alias{seed_sequence<-}
\title{Set miRNA Seed Sequence}
\usage{
seed_sequence(object) <- value
}
\arguments{
\item{object}{Object of class \code{miRNAGene}.}

\item{value}{New value for seed_sequence.}
}
\value{
The modified object.
}
\description{
Setter for seed_sequence attribute of \code{miRNAGene} class.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
mir1 <- miRNAGene(seed_sequence = 'AUUA')
seed_sequence(mir1)
seed_sequence(mir1) <- 'AAAGCA'
seed_sequence(mir1)
}
\seealso{
\link{seed_sequence<-,miRNAGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
