% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_siRNAGene.R
\docType{class}
\name{siRNAGene-class}
\alias{siRNAGene-class}
\title{siRNAGene Class}
\value{
An object of class \code{siRNAGene}.
}
\description{
Class to represent genes encoding for short interfering RNAs.
}
\section{Slots}{

\describe{
\item{\code{siRNA_mature_sequence}}{\code{RNAString}. siRNA mature sequence.}

\item{\code{siRNA_off_targets}}{\code{list}. siRNA off-targets.}
}}

\examples{
# An empty siRNAGene object
sir1 <- new("siRNAGene")
}
\seealso{
\linkS4class{RegulatoryNonCodingGene}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
