% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_siRNAGene.R
\name{siRNAGene-validity}
\alias{siRNAGene-validity}
\title{Validity Check for siRNAGene Class}
\arguments{
\item{object}{Object of class \code{siRNAGene}.}
}
\value{
\code{TRUE} if the object is valid, otherwise the function
throws an error via \code{stop()}.
}
\description{
The function checks that mature siRNA sequence is consistent
with the gene sequence and coordinates defined.
}
\seealso{
\link{RegulatoryNonCodingGene-validity}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
