% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_SmallNonCodingGene.R
\docType{class}
\name{SmallNonCodingGene-class}
\alias{SmallNonCodingGene-class}
\title{SmallNonCodingGene Virtual Class}
\value{
No objects can be created directly from this class since
it's virtual.
}
\description{
This virtual class is the class from which
miRNAGene, siRNAGene, piRNAGene, snRNAGene and snoRNAGene classes are
derived.
The class has no additional slots w.r.t. RegulatoryNonCodingGene class
because it was created just for the sake of clarity.
}
\examples{
# The SmallNonCodingGene class is virtual, so no objects can
# be created directly from this class. It needs to be extended by
# other classes.
}
\seealso{
\linkS4class{RegulatoryNonCodingGene}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
