% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_snoRNAGene.R
\docType{class}
\name{snoRNAGene-class}
\alias{snoRNAGene-class}
\title{snoRNAGene Class}
\value{
An object of class \code{snoRNAGene}.
}
\description{
Class to represent genes encoding for small nucleolar RNAs.
}
\section{Slots}{

\describe{
\item{\code{snoRNA_mature_sequence}}{\code{RNAString}. snoRNA mature sequence.}

\item{\code{snoRNA_box_type}}{\code{character}. snoRNA's box type.}
}}

\examples{
# An empty snoRNAGene object
sno1 <- new("snoRNAGene")
}
\seealso{
\linkS4class{RegulatoryNonCodingGene}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
