% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_snoRNAGene.R
\name{snoRNAGene}
\alias{snoRNAGene}
\title{Create an Object of Class snoRNAGene}
\usage{
snoRNAGene(
    gene_ensembl_id = "ENSG00000000000",
    hugo_symbol = NA_character_,
    gene_complete_name = NA_character_,
    gene_description = NA_character_,
    strand = NA,
    chromosome = NA,
    gene_start = NA,
    gene_end = NA,
    gene_sequence = DNAString(),
    exons_starts = NA,
    exons_ends = NA,
    pre_ncRNA_ensembl_id = "ENST00000000000",
    pre_ncRNA_start = NA,
    pre_ncRNA_end = NA,
    regulatory_functions = list(),
    targets_ensembl_id = list(),
    snoRNA_mature_sequence = RNAString(),
    snoRNA_box_type = NA_character_ )
}
\arguments{
\item{gene_ensembl_id}{Gene ENSEMBL ID.}

\item{hugo_symbol}{Gene Hugo symbol.}

\item{gene_complete_name}{Gene complete name.}

\item{gene_description}{Gene description.}

\item{strand}{Strand on which the gene is annotated.}

\item{chromosome}{Chromosome on which the gene is annotated.}

\item{gene_start}{Gene start coordinate.}

\item{gene_end}{Gene end coordinate.}

\item{gene_sequence}{Gene sequence.}

\item{exons_starts}{Exons starts coordinates.}

\item{exons_ends}{Exons ends coordinates.}

\item{pre_ncRNA_ensembl_id}{pre-ncRNA ENSEMBL ID.}

\item{pre_ncRNA_start}{pre-ncRNA start coordinate.}

\item{pre_ncRNA_end}{pre-ncRNA end coordinate.}

\item{regulatory_functions}{Gene regulatory functions.}

\item{targets_ensembl_id}{ENSEMBL IDs of the gene targets.}

\item{snoRNA_mature_sequence}{Mature snoRNA sequence.}

\item{snoRNA_box_type}{snoRNA box type.}
}
\value{
New snoRNAGene object.
}
\description{
Constructor function for \code{snoRNAGene} class.
}
\examples{
# An empty snoRNAGene object
sno1 <- snoRNAGene()
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
