% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_snRNAGene.R
\docType{class}
\name{snRNAGene-class}
\alias{snRNAGene-class}
\title{snRNAGene Class}
\value{
An object of class \code{snRNAGene}.
}
\description{
Class to represent genes encoding for small nuclear RNAs.
}
\section{Slots}{

\describe{
\item{\code{snRNA_mature_sequence}}{\code{RNAString}. snRNA mature sequence.}

\item{\code{spliceosome_complex}}{\code{character}.
Spliceosome complex to which snRNA is associated.}
}}

\examples{
# An empty snRNAGene object
sn1 <- new("snRNAGene")
}
\seealso{
\linkS4class{RegulatoryNonCodingGene}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
