% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_snRNAGene.R
\name{spliceosome_complex<-,snRNAGene-method}
\alias{spliceosome_complex<-,snRNAGene-method}
\title{Set snRNA Spliceosome Complex}
\usage{
\S4method{spliceosome_complex}{snRNAGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{snRNAGene}.}

\item{value}{New value for spliceosome_complex.}
}
\value{
The modified object.
}
\description{
Setter for spliceosome_complex attribute of \code{snRNAGene} class.
}
\examples{
sn1 <- snRNAGene(spliceosome_complex = 'U4')
spliceosome_complex(sn1)
spliceosome_complex(sn1) <- 'U6'
spliceosome_complex(sn1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
