% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_snRNAGene.R
\name{spliceosome_complex<-}
\alias{spliceosome_complex<-}
\title{Set snRNA Spliceosome Complex}
\usage{
spliceosome_complex(object) <- value
}
\arguments{
\item{object}{Object of class \code{snRNAGene}.}

\item{value}{New value for spliceosome_complex.}
}
\value{
The modified object.
}
\description{
Setter for spliceosome_complex attribute of \code{snRNAGene} class.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
sn1 <- snRNAGene(spliceosome_complex = 'U4')
spliceosome_complex(sn1)
spliceosome_complex(sn1) <- 'U6'
spliceosome_complex(sn1)
}
\seealso{
\link{spliceosome_complex<-,snRNAGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
