% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_snRNAGene.R
\name{spliceosome_complex,snRNAGene-method}
\alias{spliceosome_complex,snRNAGene-method}
\title{Get snRNA Spliceosome Complex}
\usage{
\S4method{spliceosome_complex}{snRNAGene}(object)
}
\arguments{
\item{object}{Object of class \code{snRNAGene}.}
}
\value{
Mature snRNA sequence.
}
\description{
Getter for spliceosome_complex attribute of \code{snRNAGene} class.
}
\examples{
sn1 <- snRNAGene(spliceosome_complex = 'U4')
spliceosome_complex(sn1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
