% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_snRNAGene.R
\name{spliceosome_complex}
\alias{spliceosome_complex}
\title{Get snRNA Spliceosome Complex}
\usage{
spliceosome_complex(object)
}
\arguments{
\item{object}{Object of class \code{snRNAGene}.}
}
\value{
Mature snRNA sequence.
}
\description{
Getter for spliceosome_complex attribute of \code{snRNAGene} class.
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
sn1 <- snRNAGene(spliceosome_complex = 'U4')
spliceosome_complex(sn1)
}
\seealso{
\link{spliceosome_complex,snRNAGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
