% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_RegulatoryNonCodingGene.R
\name{targets_id,RegulatoryNonCodingGene-method}
\alias{targets_id,RegulatoryNonCodingGene-method}
\title{Get Targets}
\usage{
\S4method{targets_id}{RegulatoryNonCodingGene}(object)
}
\arguments{
\item{object}{Object of class \code{RegulatoryNonCodingGene}
(and derived classes).}
}
\value{
List of targets ENSEMBL IDs.
}
\description{
Getter for targets_ensembl_id attribute of
\code{RegulatoryNonCodingGene} class (and derived classes).
}
\examples{
# Since RegulatoryNonCodingGene is a virtual class, the example is made
# using a derived class
mir1 <- miRNAGene(targets_ensembl_id = list("ENST12341234090"))
targets_id(mir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
