% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_RegulatoryNonCodingGene.R
\name{targets_id<-,RegulatoryNonCodingGene-method}
\alias{targets_id<-,RegulatoryNonCodingGene-method}
\title{Add/Remove Targets}
\usage{
\S4method{targets_id}{RegulatoryNonCodingGene}(object, action = "add") <- value
}
\arguments{
\item{object}{Object of class \code{RegulatoryNonCodingGene}
(and derived classes).}

\item{action}{Action to be done: can be 'add' or 'remove'.}

\item{value}{Target ENSEMBL ID to add/remove.}
}
\value{
The modified object.
}
\description{
Setter for targets_ensembl_id attribute of
\code{RegulatoryNonCodingGene} class (and derived classes). By choosing an
action, which can be 'add' or 'remove', it's possible to add and remove,
respectively, targets ENSEMBL IDs from targets_ensembl_id attribute.
}
\examples{
# Since RegulatoryNonCodingGene is a virtual class, the example is made
# using a derived class
mir1 <- miRNAGene()
targets_id(mir1)
targets_id(mir1, action = 'add') <- 'ENST09090909091'
targets_id(mir1)
targets_id(mir1, action = 'remove') <- 'ENST09090909091'
targets_id(mir1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
