% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_tRNAGene.R
\docType{class}
\name{tRNAGene-class}
\alias{tRNAGene-class}
\title{tRNAGene Class}
\value{
An object of class \code{tRNAGene}.
}
\description{
Class to represent genes encoding for transfer RNAs.
}
\section{Slots}{

\describe{
\item{\code{tRNA_mature_sequence}}{\code{RNAString}. tRNA mature sequence.}

\item{\code{anticodon}}{\code{RNAString}. tRNA anticodon.}

\item{\code{amino_acid}}{\code{AAString}. tRNA aminoacid.}
}}

\examples{
# An empty tRNAGene object
t1 <- new("tRNAGene")
}
\seealso{
\linkS4class{HousekeepingNonCodingGene}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
