% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_tRNAGene.R
\name{tRNAGene-validity}
\alias{tRNAGene-validity}
\title{Validity Check for tRNAGene Class}
\arguments{
\item{object}{Object of class \code{tRNAGene}.}
}
\value{
\code{TRUE} if the object is valid, otherwise the function
throws an error via \code{stop()}.
}
\description{
The function checks that: each slot, except the special ones,
is of length 1 (or empty); anticodon and aminoacid are consistent one with
each other and that maure tRNA sequence is compatible with both the
defined coordinates and the gene sequence.
}
\seealso{
\link{HousekeepingNonCodingGene-validity}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
