% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_HousekeepingNonCodingGene.R
\name{ubiquitous_expression<-,HousekeepingNonCodingGene-method}
\alias{ubiquitous_expression<-,HousekeepingNonCodingGene-method}
\title{Set Ubiquitous Expression}
\usage{
\S4method{ubiquitous_expression}{HousekeepingNonCodingGene}(object) <- value
}
\arguments{
\item{object}{Object of class \code{HousekeepingNonCodingGene}
(and derived classes).}

\item{value}{New value for ubiquitous_expression.}
}
\value{
The modified object.
}
\description{
Setter for ubiquitous_expression attribute of
\code{HousekeepingNonCodingGene} class (and derived classes).
}
\examples{
# Since HousekeepingNonCodingGene is a virtual class, the example is made
# using a derived class
t1 <- tRNAGene(ubiquitous_expression = FALSE)
ubiquitous_expression(t1)
ubiquitous_expression(t1) <- TRUE
ubiquitous_expression(t1)
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
