% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_HousekeepingNonCodingGene.R
\name{ubiquitous_expression}
\alias{ubiquitous_expression}
\title{Get Ubiquitous Expression}
\usage{
ubiquitous_expression(object)
}
\arguments{
\item{object}{Object of class \code{HousekeepingNonCodingGene}
(and derived classes).}
}
\value{
Value of ubiquitous_expression.
}
\description{
Getter for ubiquitous_expression attribute of
\code{HousekeepingNonCodingGene} class (and derived classes).
}
\examples{
# This is only the generic function, look in 'See Also' section
# for the method implementation.
t1 <- tRNAGene(ubiquitous_expression = FALSE)
ubiquitous_expression(t1)
}
\seealso{
\link{ubiquitous_expression,HousekeepingNonCodingGene-method}
}
\author{
Marco Cominelli, \email{marco.cominelli@mail.polimi.it}
}
